Add PNG8 "compression" support
PNG8 images are limited to 256 colors, and can be much smaller files than similar PNG32 images. 45% reductions on some complex low-color atlases are the low end, and some more-lossy conversions on high-color atlases can reduce the size further. Here, images are more often limited to 255 opaque colors because of some limitations to allow transparent pixels to be treated specially. If you want, I can expose more of PaletteReducer in the application somehow, so the user could specify a precise palette somehow instead of the app counting color frequency and using the threshold parameter to determine what should be in the palette.

Some of the translations probably aren't correct. I used Google Translate for some, but I'm pretty sure German is mostly OK since the terms are from Wikipedia's German edition ("Farbpaletten-Modus" is definitely "color palette mode"). Russian and Traditional Chinese are not possible for me to verify on my own.

This also corrects a typo in the README.