Update PNG8 and PaletteReducer
These have seen some attention in my code since my earlier PR, and I fixed an issue with the last version where it could omit a color from a palette because it occurred the same number of times in an image as another color. I've checked and this does run fine, though I needed to adjust the project setup to be able to build (that will be in another commit, feel free to ignore it).